Move Service protocol to Service.swift
